Revert "Arrow writing"
