+++ dataframe-arrow/build.gradle.kts
@@ -14,7 +14,6 @@ dependencies {
     implementation(libs.arrow.memory)
     implementation(libs.commonsCompress)
     implementation(libs.kotlin.reflect)
-    implementation(libs.kotlin.datetimeJvm)
 
     testApi(project(":core"))
     testImplementation(libs.junit)
