+++ dataframe-arrow/src/main/kotlin/org/jetbrains/kotlinx/dataframe/io/arrow.kt
@@ -37,9 +37,9 @@ import org.jetbrains.kotlinx.dataframe.AnyFrame
 import org.jetbrains.kotlinx.dataframe.DataColumn
 import org.jetbrains.kotlinx.dataframe.DataFrame
 import org.jetbrains.kotlinx.dataframe.api.Infer
-import org.jetbrains.kotlinx.dataframe.api.NullabilityException
 import org.jetbrains.kotlinx.dataframe.api.NullabilityOptions
 import org.jetbrains.kotlinx.dataframe.api.applyNullability
+import org.jetbrains.kotlinx.dataframe.api.NullabilityException
 import org.jetbrains.kotlinx.dataframe.api.cast
 import org.jetbrains.kotlinx.dataframe.api.dataFrameOf
 import org.jetbrains.kotlinx.dataframe.api.emptyDataFrame
@@ -111,11 +111,7 @@ internal fun <T> Iterable<DataFrame<T>>.concatKeepingSchema(): DataFrame<T> {
 /**
  * Read [Arrow interprocess streaming format](https://arrow.apache.org/docs/java/ipc.html#writing-and-reading-streaming-format) data from existing [channel]
  */
-public fun DataFrame.Companion.readArrowIPC(
-    channel: ReadableByteChannel,
-    allocator: RootAllocator = Allocator.ROOT,
-    nullability: NullabilityOptions = NullabilityOptions.Infer
-): AnyFrame {
+public fun DataFrame.Companion.readArrowIPC(channel: ReadableByteChannel, allocator: RootAllocator = Allocator.ROOT, nullability: NullabilityOptions = NullabilityOptions.Infer): AnyFrame {
     ArrowStreamReader(channel, allocator).use { reader ->
         val dfs = buildList {
             val root = reader.vectorSchemaRoot
@@ -132,11 +128,7 @@ public fun DataFrame.Companion.readArrowIPC(
 /**
  * Read [Arrow random access format](https://arrow.apache.org/docs/java/ipc.html#writing-and-reading-random-access-files) data from existing [channel]
  */
-public fun DataFrame.Companion.readArrowFeather(
-    channel: SeekableByteChannel,
-    allocator: RootAllocator = Allocator.ROOT,
-    nullability: NullabilityOptions = NullabilityOptions.Infer
-): AnyFrame {
+public fun DataFrame.Companion.readArrowFeather(channel: SeekableByteChannel, allocator: RootAllocator = Allocator.ROOT, nullability: NullabilityOptions = NullabilityOptions.Infer): AnyFrame {
     ArrowFileReader(channel, allocator).use { reader ->
         val dfs = buildList {
             reader.recordBlocks.forEach { block ->
@@ -171,9 +163,8 @@ private fun Float8Vector.values(range: IntRange): List<Double?> = range.map { ge
 
 private fun DurationVector.values(range: IntRange): List<Duration?> = range.map { getObject(it) }
 private fun DateDayVector.values(range: IntRange): List<LocalDate?> = range.map {
-    if (getObject(it) == null) null else {
-        DateUtility.getLocalDateTimeFromEpochMilli(getObject(it).toLong() * DateUtility.daysToStandardMillis).toLocalDate()
-    }
+    if (getObject(it) == null) null else
+    DateUtility.getLocalDateTimeFromEpochMilli(getObject(it).toLong() * DateUtility.daysToStandardMillis).toLocalDate()
 }
 private fun DateMilliVector.values(range: IntRange): List<LocalDateTime?> = range.map { getObject(it) }
 
@@ -198,12 +189,9 @@ private fun TimeMilliVector.values(range: IntRange): List<LocalTime?> = range.ma
         LocalTime.ofNanoOfDay(get(it).toLong() * 1000_000)
     }
 }
-private fun TimeSecVector.values(range: IntRange): List<LocalTime?> = range.map {
-    getObject(it)?.let { LocalTime.ofSecondOfDay(it.toLong()) }
-}
-private fun StructVector.values(range: IntRange): List<Map<String, Any?>?> = range.map {
-    getObject(it)
-}
+private fun TimeSecVector.values(range: IntRange): List<LocalTime?> = range.map { getObject(it)?.let {LocalTime.ofSecondOfDay(it.toLong())}  }
+
+private fun StructVector.values(range: IntRange): List<Map<String, Any?>?> = range.map { getObject(it) }
 
 private fun VarCharVector.values(range: IntRange): List<String?> = range.map {
     if (isNull(it)) {
@@ -237,10 +225,7 @@ private fun LargeVarCharVector.values(range: IntRange): List<String?> = range.ma
     }
 }
 
-private inline fun <reified T> List<T?>.withTypeNullable(
-    expectedNulls: Boolean,
-    nullabilityOptions: NullabilityOptions
-): Pair<List<T?>, KType> {
+private inline fun <reified T> List<T?>.withTypeNullable(expectedNulls: Boolean, nullabilityOptions: NullabilityOptions): Pair<List<T?>, KType> {
     val nullable = nullabilityOptions.applyNullability(this, expectedNulls)
     return this to typeOf<T>().withNullability(nullable)
 }
@@ -289,34 +274,25 @@ private fun readField(root: VectorSchemaRoot, field: Field, nullability: Nullabi
 /**
  * Read [Arrow interprocess streaming format](https://arrow.apache.org/docs/java/ipc.html#writing-and-reading-streaming-format) data from existing [file]
  */
-public fun DataFrame.Companion.readArrowIPC(
-    file: File,
-    nullability: NullabilityOptions = NullabilityOptions.Infer
-): AnyFrame = Files.newByteChannel(file.toPath()).use { readArrowIPC(it, nullability = nullability) }
+public fun DataFrame.Companion.readArrowIPC(file: File, nullability: NullabilityOptions = NullabilityOptions.Infer): AnyFrame =
+        Files.newByteChannel(file.toPath()).use { readArrowIPC(it, nullability = nullability) }
 
 /**
  * Read [Arrow interprocess streaming format](https://arrow.apache.org/docs/java/ipc.html#writing-and-reading-streaming-format) data from existing [byteArray]
  */
-public fun DataFrame.Companion.readArrowIPC(
-    byteArray: ByteArray,
-    nullability: NullabilityOptions = NullabilityOptions.Infer
-): AnyFrame = SeekableInMemoryByteChannel(byteArray).use { readArrowIPC(it, nullability = nullability) }
+public fun DataFrame.Companion.readArrowIPC(byteArray: ByteArray, nullability: NullabilityOptions = NullabilityOptions.Infer): AnyFrame =
+        SeekableInMemoryByteChannel(byteArray).use { readArrowIPC(it, nullability = nullability) }
 
 /**
  * Read [Arrow interprocess streaming format](https://arrow.apache.org/docs/java/ipc.html#writing-and-reading-streaming-format) data from existing [stream]
  */
-public fun DataFrame.Companion.readArrowIPC(
-    stream: InputStream,
-    nullability: NullabilityOptions = NullabilityOptions.Infer
-): AnyFrame = Channels.newChannel(stream).use { readArrowIPC(it, nullability = nullability) }
+public fun DataFrame.Companion.readArrowIPC(stream: InputStream, nullability: NullabilityOptions = NullabilityOptions.Infer): AnyFrame =
+        Channels.newChannel(stream).use { readArrowIPC(it, nullability = nullability) }
 
 /**
  * Read [Arrow interprocess streaming format](https://arrow.apache.org/docs/java/ipc.html#writing-and-reading-streaming-format) data from existing [url]
  */
-public fun DataFrame.Companion.readArrowIPC(
-    url: URL,
-    nullability: NullabilityOptions = NullabilityOptions.Infer
-): AnyFrame =
+public fun DataFrame.Companion.readArrowIPC(url: URL, nullability: NullabilityOptions = NullabilityOptions.Infer): AnyFrame =
     when {
         isFile(url) -> readArrowIPC(urlAsFile(url), nullability)
         isProtocolSupported(url) -> url.openStream().use { readArrowIPC(it, nullability) }
@@ -325,10 +301,7 @@ public fun DataFrame.Companion.readArrowIPC(
         }
     }
 
-public fun DataFrame.Companion.readArrowIPC(
-    path: String,
-    nullability: NullabilityOptions = NullabilityOptions.Infer
-): AnyFrame = if (isURL(path)) {
+public fun DataFrame.Companion.readArrowIPC(path: String, nullability: NullabilityOptions = NullabilityOptions.Infer): AnyFrame = if (isURL(path)) {
     readArrowIPC(URL(path), nullability)
 } else {
     readArrowIPC(File(path), nullability)
@@ -339,34 +312,25 @@ public fun DataFrame.Companion.readArrowIPC(
 /**
  * Read [Arrow random access format](https://arrow.apache.org/docs/java/ipc.html#writing-and-reading-random-access-files) data from existing [file]
  */
-public fun DataFrame.Companion.readArrowFeather(
-    file: File,
-    nullability: NullabilityOptions = NullabilityOptions.Infer
-): AnyFrame = Files.newByteChannel(file.toPath()).use { readArrowFeather(it, nullability = nullability) }
+public fun DataFrame.Companion.readArrowFeather(file: File, nullability: NullabilityOptions = NullabilityOptions.Infer): AnyFrame =
+        Files.newByteChannel(file.toPath()).use { readArrowFeather(it, nullability = nullability) }
 
 /**
  * Read [Arrow random access format](https://arrow.apache.org/docs/java/ipc.html#writing-and-reading-random-access-files) data from existing [byteArray]
  */
-public fun DataFrame.Companion.readArrowFeather(
-    byteArray: ByteArray,
-    nullability: NullabilityOptions = NullabilityOptions.Infer
-): AnyFrame = SeekableInMemoryByteChannel(byteArray).use { readArrowFeather(it, nullability = nullability) }
+public fun DataFrame.Companion.readArrowFeather(byteArray: ByteArray, nullability: NullabilityOptions = NullabilityOptions.Infer): AnyFrame =
+        SeekableInMemoryByteChannel(byteArray).use { readArrowFeather(it, nullability = nullability) }
 
 /**
  * Read [Arrow random access format](https://arrow.apache.org/docs/java/ipc.html#writing-and-reading-random-access-files) data from existing [stream]
  */
-public fun DataFrame.Companion.readArrowFeather(
-    stream: InputStream,
-    nullability: NullabilityOptions = NullabilityOptions.Infer
-): AnyFrame = readArrowFeather(stream.readBytes(), nullability)
+public fun DataFrame.Companion.readArrowFeather(stream: InputStream, nullability: NullabilityOptions = NullabilityOptions.Infer): AnyFrame =
+        readArrowFeather(stream.readBytes(), nullability)
 
 /**
  * Read [Arrow random access format](https://arrow.apache.org/docs/java/ipc.html#writing-and-reading-random-access-files) data from existing [url]
  */
-public fun DataFrame.Companion.readArrowFeather(
-    url: URL,
-    nullability: NullabilityOptions = NullabilityOptions.Infer
-): AnyFrame =
+public fun DataFrame.Companion.readArrowFeather(url: URL, nullability: NullabilityOptions = NullabilityOptions.Infer): AnyFrame =
     when {
         isFile(url) -> readArrowFeather(urlAsFile(url), nullability)
         isProtocolSupported(url) -> readArrowFeather(url.readBytes(), nullability)
@@ -378,10 +342,7 @@ public fun DataFrame.Companion.readArrowFeather(
 /**
  * Read [Arrow random access format](https://arrow.apache.org/docs/java/ipc.html#writing-and-reading-random-access-files) data from existing [path]
  */
-public fun DataFrame.Companion.readArrowFeather(
-    path: String,
-    nullability: NullabilityOptions = NullabilityOptions.Infer
-): AnyFrame = if (isURL(path)) {
+public fun DataFrame.Companion.readArrowFeather(path: String, nullability: NullabilityOptions = NullabilityOptions.Infer): AnyFrame = if (isURL(path)) {
     readArrowFeather(URL(path), nullability)
 } else {
     readArrowFeather(File(path), nullability)
