+++ docs/StardustDocs/topics/read.md
@@ -193,9 +193,8 @@ implementation("org.jetbrains.kotlinx:dataframe-arrow:$dataframe_version")
 Make sure to follow [Apache Arrow Java compatibility](https://arrow.apache.org/docs/java/install.html#java-compatibility) guide when using Java 9+ 
 </warning>
 
-Dataframe supports reading [Arrow interprocess streaming format](https://arrow.apache.org/docs/java/ipc.html#writing-and-reading-streaming-format)
-and [Arrow random access format](https://arrow.apache.org/docs/java/ipc.html#writing-and-reading-random-access-files)
-from raw Channel (ReadableByteChannel for streaming and SeekableByteChannel for random access), InputStream, File or ByteArray.
+Dataframe supports reading from [Arrow interprocess streaming format](https://arrow.apache.org/docs/java/ipc.html#writing-and-reading-streaming-format) and [Arrow random access format](https://arrow.apache.org/docs/java/ipc.html#writing-and-reading-random-access-files)
+
 <!---FUN readArrowFeather-->
 
 ```kotlin
@@ -203,3 +202,5 @@ val df = DataFrame.readArrowFeather(file)
 ```
 
 <!---END-->
+
+
